requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.6
django-filters==0.2.1
djangorestframework==3.11.0
drf-nested-routers==0.91
gunicorn==20.0.4
psycopg2-binary
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
Pillow==7.1.2

